inference/requirements.txt
peft>=0.11.0
Pillow>=9.5.0